refactor: move messageCreate event to the proper place
